+++ CHANGELOG.md
@@ -8,3 +8,7 @@
 - Explicitly add a submit button for adding new tasks
 - Improve `BetaPage` component with title, subtitle, and actions
 - Import AccountID from URL parameter (step toward sharing the account with another device)
+- Improve layouting and Account guarding
+- Share account via QR Code
+- Copy share url to clipboard
+- Share project via QR Code and Clipboard